fix(notifications): Fix bug where delete notification may result in a client-side error
When a user deletes a resource, it triggers a notification of a delete.  The user who intiated the

delete will receive this notificaiton.  The way a delete notification was handled was to call

`record.deleteRecord()`, which is incorrect in this scenario.  This is resolved by calling

`record.unloadRecord()` instead, which is expected to work both when a record was deleted by another

user or the current user.  Fixes [#137728433]