fix(notifications): Fixed bug where delete notification handled while delete request is in-flight ca
Attempting to unload a record that is in-flight causes an error, which happens when a delete

notification is received so quickly as to be handled before a delete request is complete.  This is

resolved in the application router by checking if the record `isSaving` and waiting for completion

if so.  Fixes [#137728433]