fix(auth): Redirect to standard post-auth route before logout
Normally, the login form attemps to redirect the user to the last seen route.  When a user logs out
while on a page inaccessible to another user who subsequently logs in, this results in an error.
This fix redirects to the standard post-auth route, accessible to all users, before logging out.

Fixes [#139828299]